Replace BackgroundJob → Reporter.RunInBackground
Turns out that BackgroundJob really wasn't the right abstraction; among
other things, it had a race when used in a more general case.
